+++ tests/test_python_server.py
@@ -75,3 +75,25 @@ def test_logging(tmp_path):
     assert validate_ping.validate_ping(input, output, schema_url=schema_url) == 0, (
         output.getvalue()
     )
+
+
+def test_parser_python_server_metrics_unsupported_type(tmp_path, capsys):
+    """Test that no files are generated with unsupported metric types."""
+    translate.translate(
+        [
+            ROOT / "data" / "python_server_metrics_unsupported.yaml",
+        ],
+        "python_server",
+        tmp_path,
+    )
+    captured = capsys.readouterr()
+    unsupported_types = [
+        "labeled_boolean",
+        "labeled_string",
+        "string_list",
+        "timespan",
+        "url",
+        "uuid",
+    ]
+    for t in unsupported_types:
+        assert f"Ignoring unsupported metric type: {t}" in captured.out